[Feat] 공격 코드 수정

+++ Project_SphereWarrior/Assets/001.Scripts/Managers/ObjectManager.cs
@@ -28,7 +28,7 @@ public class ObjectManager
     private Transform monsterTrans;
     #endregion
 
-    public MonsterController SpawnMonster(int _index, int _hp,  Vector3 _position)
+    public MonsterController SpawnMonster(int _index, float _hp,  Vector3 _position)
     {
         MonsterController monster = Managers.Resource.Instantiate($"Monster_{_index}", MonsterTrans, true).GetComponent<MonsterController>();
         monster.transform.position = _position;
